refactor(pages) improved pages' code to handle coming features
1. Move the code from '/' to '/feed' for coming features
2. Move images to ~/assets/images
3. Remove unnecessary files
