+++ pages/chat.vue
@@ -1,6 +1,3 @@
-<script setup>
-  import myImage from "./Omar.jpeg"
-</script>
 <template>
   <main class="relative grid grid-cols-1 items-center justify-center bg-green-50 w-screen min-h-screen pl-[100px] overflow-x-hidden">
     <Sidebar :profile-image="myImage" />
